refactor(articlelist-container-styles): move styles for articlelist container to new folder
